--- src/maison/readers/__init__.py
@@ -1,6 +1,0 @@
-from .ini import IniReader
-from .pyproject import PyprojectReader
-from .toml import TomlReader
-
-
-__all__ = ["IniReader", "PyprojectReader", "TomlReader"]
